refactor: Move parser pkg to pkg
