+++ cmd/convert.go
@@ -9,7 +9,7 @@ import (
 	"path"
 
 	"github.com/glass-cms/glasscms/internal/item"
-	"github.com/glass-cms/glasscms/parser"
+	"github.com/glass-cms/glasscms/pkg/parser"
 	"github.com/glass-cms/glasscms/pkg/sourcer"
 	"github.com/lmittmann/tint"
 	"github.com/spf13/cobra"
