+++ pkg/parser/parser_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/glass-cms/glasscms/parser"
+	"github.com/glass-cms/glasscms/pkg/parser"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
